// webapp/src/api/model/craftingRecipe.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { CraftingRecipeMaterialsItem } from './craftingRecipeMaterialsItem';
import type { ItemUsage } from './itemUsage';

export interface CraftingRecipe {
  id?: string;
  materials: CraftingRecipeMaterialsItem[];
  otherCircumstances: string;
  products: ItemUsage[];
  profession: string;
  requirement: string;
}