// webapp/src/api/model/item.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { ItemEffectsItem } from './itemEffectsItem';
import type { ERarity } from './eRarity';

export interface Item {
  description: string;
  effects: ItemEffectsItem[];
  id?: string;
  maximumStackSize: number;
  minimumStackSize: number;
  name: string;
  note: string;
  rarity: ERarity;
  requirement: string;
  tags: string[];
  tier: number;
  upgradeSlots: number;
  vendorPrice: number;
}